feat(js-client): RedWire-over-WSS browser transport (#937)
Let the browser SPA talk RedWire directly over a binary WebSocket, reusing
the existing codec and swapping only the byte transport (PRD #930).

- Extract the transport-agnostic codec into `redwire-core.js` (zero `node:`
  imports; zstd is injected via `setZstdProvider` rather than importing
  `node:zlib`). `redwire.js` becomes a thin Node shim adding the TCP/TLS
  socket openers + node zstd, and `connectRedwire` now accepts an injected
  `socket`. Re-exports keep existing importers working.
- `redwire-ws.js`: `WebSocketDuplex` adapts a binary WebSocket to the
  node-socket-shaped duplex the codec consumes; `connectRedwireWs` runs the
  standard handshake over it.
- `red+wss://host:port` scheme parses to a `wss://host:port/redwire`
  endpoint and is wired into the browser `connect()`; `.d.ts` updated.

Tests: new `redwire-ws.test.mjs` proves a handshake + multiplexed query
round-trip and frame encode/decode parity over a mock binary WebSocket
(the mock server reuses the same codec). Full JS suite 165 pass / 2 live-
skipped, including the browser-portability guard (no `node:` in the bundle
graph; `redwire.js` stays out — only the node-free core enters).
