+++ drivers/js-client/src/index.browser.js
@@ -40,6 +40,7 @@ import {
   mergeAuthFromUri,
 } from './core/index.js'
 import { HttpRpcClient } from './http.js'
+import { connectRedwireWs, REDWIRE_WS_PATH } from './redwire-ws.js'
 import { createSelectStream, createInputStream } from './streaming-web.js'
 
 export { RedDBError, EmbeddedNotSupported, EMBEDDED_REJECTION_MESSAGE, isEmbeddedUri }
@@ -119,6 +120,28 @@ export async function connect(uri, options = {}) {
     }
     const client = new HttpRpcClient({ baseUrl, token })
     await client.call('query', { sql: 'SELECT 1' })
+    return new RedDB(client)
+  }
+
+  // RedWire-over-binary-WebSocket (#937, ADR 0036): the browser speaks the
+  // same multiplexed binary protocol as the native drivers, tunneled over
+  // a WSS the sandbox can open. The TLS edge enforces the Origin allowlist
+  // and WSS-only on its side; here we just open the socket and run the
+  // standard RedWire handshake over it.
+  if (parsed.kind === 'redwss') {
+    const merged = mergeAuthFromUri(parsed, options.auth)
+    let token = merged.token
+    if (!token && merged.username && merged.password) {
+      const loginUrl = merged.loginUrl ?? `https://${parsed.host}:${parsed.port}/auth/login`
+      const session = await login(loginUrl, {
+        username: merged.username,
+        password: merged.password,
+      })
+      token = session.token
+    }
+    const auth = token ? { kind: 'bearer', token } : { kind: 'anonymous' }
+    const url = `wss://${parsed.host}:${parsed.port}${REDWIRE_WS_PATH}`
+    const client = await connectRedwireWs({ url, auth })
     return new RedDB(client)
   }
